Allow HTTPS NBC Learn video and fix warning.

+++ ImportHyperlinkPlugin.php
@@ -255,8 +255,8 @@ class ImportHyperlinkPlugin extends Omeka_Plugin_AbstractPlugin
             }
         }
 
-        // Check that the URL was not a "link" type, and an image is available.
-        if ($embed->type != 'link' && $embed->image) {
+        // Check that the URL was not a "link" type, and an image is available.  
+        if (!empty($embed) && $embed->type != 'link' && $embed->image) {
             // Delete all existing files for the item.
             foreach ($record->getFiles() as $file) {
                 $file->delete();
